replace stock strategy system
The stock strategies are so penalizing that it's universally agreed that there are nearly no cases where it's actually advantageous to take a single one.

They could be fixed with simple rebalancing to make them less unbalanced. However, no mod currently exists on CKAN that does that and supports versions later than 0.90.

Strategia is a complete rewrite of the strategy system. Normally that would disqualify it from a stockalike modpack. However - since the stock strategies are completely unusable anyway, this is a far better option. At worst you can completely ignore it, just like stock.